Test random position generation for given depth
I used CorticalColumn to generate random positions by depth.
This implementation works only for circuits with an atlas ---
with "[PH]y" data. CircuitGeometry makes a call to get
CorticalColumn for given region. Class names might suggest
that this case works only with cortical regions, though
that is not the case. I have used the terms CorticalDepth and
CorticalColumn as I was working with cortical circuits, and could
not think of better names. These classes should work with any atlas
based circuit that provides a "depth" into the circuit via "[PH]y".