Add measurement for excitatory synaptic density
So far we had a 'CompositionMeasurementMethod' for
extrinsic + intrinsic synapse density that multiplied
segment length in the ROI with a known spine density
to get an estimate of synapse density. This method
was used to compute synapses between cells in the ROI,
as well as those with an incoming projection. The former
(intrinsic synapses) can be counted directly using the
circuit's connectome's spatial index.