Implement bar plot for an analysis
The previous implementation of the analysis bar plot was
a left over from earlier iterations when this code was used
for validation comparison plots. Validation comparison plots
are now made using the ComparisonPlot subclass BarPlot.
The new analysis bar plot should be used for analysis that
a model measurement containing a dataframe with a simple index,
and not a pandas.MultiIndex.