Test develop circuit connectome analysis
The BrainCircuitMeasurementParameter for the connectome
depends on the circuit instance for its values. As a result
we cannot set it before we know the circuit that will be analyzed.
We provide a method to these parameters that accepts a circuit and
updates its values accordingly. This method may be used once the
circuit to be analyzed is known. Otherwise, we allow these parameters
to be defined without the circuit --- if the circuit is provided, the
values will be validated, otherwise they will remain unvalidated.

1. Generalize and Move ByLayerValidationTestCase up as a
   CircuitPhenomenonValidationTestCase
2. Move data to composition/xxx_data, so that we can have
   connectome/xxx_data
3. Generalize and move Blue Brain Data subclasses up. Data
   classes were written for composition data, but their code
   can be generalized...