Plot multiple plots in an analysis
MtypeCellDensityAnalysis will now produce a report
with several figures. Each of these figures is an individual
png. A MultiPlot class has also been defined that will
produce a figure for each of the columns in a multi indexed
column dataframe.
The MultiBarPlot can be used for MtypeCellDensityAnalysis,
but that is not the priority at the moment.