determine in validation if condition is permissble
Analysis class has been proviced a method "is_permissible",
with a default return of True. In validations, this method
determines if the validation data contains the condition (as_dict)
passed to it as an argument. The method is placed as a call back
(lambda) key-word argument in method 'get_measurement' ---
which should then be forwarded to Adapter's method calls.
I also added an argument to 'fill_missing_param_values' in
BBP Adapter's statistical measurement method. A FiniteValuedParameter
was used to fill in the missing rows in a measurement --- however
with Validation's data determining which parameter value (combinations)
to measure, filling missing rows became problematic.
For now, adapter's pathway_measurements will not fill missing rows by
default, while spatial_measurements will continue to...