Topic: Connection Probabilities by Distance
Index mangling happened again --- though this
time understood what causes it: concatenating
incompatible dataframes.
The incompatible dataframe was produced when the
located index value has only one match --- thus
producing a series (analogous to a single row).
When more than a sinlgle rows match, the result is
a dataframe. I have carefully accounted for these
possibilities.
One thing that should be developed next is to
run the analysis for selected pathways. This is hard
to achieve as the pathways are determined as the cross
product of the pre and post mtypes. Only relevant
pathways can then be processed by passing in a
'is_permissible' function that will say yes only
to the desired (pre_mtype, post_mtype) pathways.

1. Specify line plot axes limits, making sure that there are
    at least two data points in the data.
2. Set colors by connection type in synapse counts.
    We should implement a general color scheme for, for example
    all pathway analyses...
3. Specify pathways to analyze as a Field in connectome analyzes,
    proving a default value of empty set.