Implement efferent connection counts
Efferent connection count is a phenomenon defined
for a (pre-synaptic) cell. If we do not care about the
mtype of the post-synapticn cell of an efferent connection, we
look at the total number of efferent connections. If we are
interested in how many efferent connections a cell makes to
a population of post-mtype cells, we aggregate efferent connections
by their post-mtype. To get a statistical estimate, we sample
several pre-synaptic cells of the same (pre-)mtype...