Move version 2 code (v2) at top dmt top level
So now both dmt, neuro_dmt will be in v2.
Once we are done refactoring, we can move
contents of v2 under top level dmt.
This will allow use to not have to write 'v2' in
our imports in the dmt version two code.
For seemless work at the shell, we add
/path/to/dmt/v2 instead of /path/to/dmt
to our environment.